+++ Modules/Intent.Modules.ModularMonolith.Host/Templates/ModuleConfiguration/ModuleConfigurationTemplatePartial.cs
@@ -34,6 +34,8 @@ namespace Intent.Modules.ModularMonolith.Host.Templates.ModuleConfiguration
                     @class.AddMethod($"IEnumerable<{this.GetModuleInstallerInterfaceName()}>", "GetModuleInstallers", method =>
                     {
                         method.Static();
+                        method.AddStatement($"var result = new List<{this.GetModuleInstallerInterfaceName()}>();");
+                        method.AddStatement($"return result;");
                     });
                 });
         }