+++ Modules/Intent.Modules.Application.MediatR.CRUD/release-notes.md
@@ -1,6 +1,7 @@
 ### Version 6.0.30
 
 - Improvement: When a create operation is mapped to Cosmos DB document with a partition key which is separate from its primary key, the auto generated primary key can now be returned.
+- Improvement: Support for different `MapTo` Extension methods.
 - Fixed: Mapping code now correctly generated when modeling and mapping a ValueObject collection as an Entity Attribute.
 
 ### Version 6.0.29