+++ Modules/Intent.Modules.MediatR.DomainEvents/release-notes.md
@@ -1,5 +1,6 @@
 ### Version 5.0.27
 
+- Improvement: Support for different `MapTo` Extension methods.
 - Fixed: Mapping code now correctly generated when modeling and mapping a ValueObject collection as an Entity Attribute.
 
 ### Version 5.0.26